requirements.txt
fastapi
uvicorn
python-dotenv
twilio
pydantic
rapidfuzz
python-multipart